convert prostate 3d example
