--- examples/prostate/prostate_3D/configs/prostate_fedavg/config/config_fed_client.json
@@ -1,34 +1,0 @@
-{
-  "format_version": 2,
-
-  "executors": [
-    {
-      "tasks": [
-        "train", "validate"
-      ],
-      "executor": {
-        "id": "Executor",
-        "name": "LearnerExecutor",
-        "args": {
-          "learner_id": "prostate-learner"
-        }
-      }
-    }
-  ],
-
-  "task_result_filters": [
-  ],
-  "task_data_filters": [
-  ],
-
-  "components": [
-    {
-      "id": "prostate-learner",
-      "path": "pt.learners.supervised_monai_prostate_learner.SupervisedMonaiProstateLearner",
-      "args": {
-        "train_config_filename": "config_train.json",
-        "aggregation_epochs": 10
-      }
-    }
-  ]
-}
