+++ examples/prostate/prostate_3D/job_configs/prostate_central/app/config/config_fed_client.json
@@ -1,0 +1,34 @@
+{
+  "format_version": 2,
+
+  "executors": [
+    {
+      "tasks": [
+        "train", "validate"
+      ],
+      "executor": {
+        "id": "Executor",
+        "name": "LearnerExecutor",
+        "args": {
+          "learner_id": "prostate-learner"
+        }
+      }
+    }
+  ],
+
+  "task_result_filters": [
+  ],
+  "task_data_filters": [
+  ],
+
+  "components": [
+    {
+      "id": "prostate-learner",
+      "path": "learners.supervised_monai_prostate_learner.SupervisedMonaiProstateLearner",
+      "args": {
+        "train_config_filename": "config_train.json",
+        "aggregation_epochs": 10
+      }
+    }
+  ]
+}
