+++ examples/prostate/prostate_3D/job_configs/prostate_ditto/app/config/config_fed_server.json
