+++ examples/prostate/prostate_3D/job_configs/prostate_fedavg/app/config/config_train.json
