+++ examples/prostate/prostate_3D/job_configs/prostate_fedavg/custom/utils/custom_client_datalist_json_path.py
