+++ examples/prostate/prostate_3D/job_configs/prostate_fedprox/app/config/config_train.json
