+++ examples/prostate/result_stat/prostate_3d_test_only.py
@@ -46,7 +46,6 @@ def main():
     dataset_base_dir = args.dataset_base_dir
     datalist_json_path = args.datalist_json_path
     model_path = args.model_path
-    cache_rate = args.cache_rate
     infer_roi_size = (224, 224, 32)
 
     device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
